polish puplisher scripts, add tests and edit change log
